apps/pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
prefect>=3.0.0
finnhub-python>=2.4.0
fastapi>=0.115.0
uvicorn>=0.32.0
